requirements.txt
datasets
ipywidgets
jupyterlab
lightgbm
matplotlib
pandas
pyyaml
scikit-learn
tqdm
transformers[torch]